feat: UTXO fetching optimisations + Evonet support fixes + Typings (#107)
* fix(Account): typings

* style(DAPIClient): adding silly level log of transporter's requests

* feat(BaseTransporter): moved executors to own property + optimisation

* chore: bump to 6.0.1

* chore: update evonet seeds

evonet seeds are now out of date

* fix: update released interval prop - subscriptions -> executors

* style: console.log -> logger.silly

* test: perform test against evonet instead of palinka

* impr: added silly log on SyncWorker / transporter announce

* impr(Worker): timeout error on failing to start

* fix(DAPIClient): correctly return set of utxos

* revert: reverted timeout on Worker

The issue is that on big wallet, it might takes more than that. This cannot be implemented this way. We need a proper implementation that takes account on potential activity of the Worker. Keeping in mind that sometimes, activity might happen while we still stalled on a action and didn't release any STARTED event

* fix: update format changed from Output to UnspendOutput

* test(coinSelection): fix Output/UTXO test

* test(strategies): deactivate test

We will rework that part soon, we need to rebuild the store with newer format

* fix: initialSyncUp event fetched_address